-define now the extra learning task for the unit operator in the world itself -restructure utils code -define train_freq as proper timedelta
