+++ assume/common/utils.py
@@ -19,6 +19,13 @@ from assume.common.market_objects import MarketProduct, Orderbook
 
 logger = logging.getLogger(__name__)
 
+freq_map = {
+    "h": rr.HOURLY,
+    "m": rr.MINUTELY,
+    "d": rr.DAILY,
+    "w": rr.WEEKLY,
+}
+
 
 def initializer(func):
     """
@@ -435,3 +442,32 @@ def timestamp2datetime(timestamp: float):
 
 def datetime2timestamp(datetime: datetime):
     return calendar.timegm(datetime.utctimetuple())
+
+
+def create_rrule(start, end, freq):
+    freq, interval = convert_to_rrule_freq(freq)
+
+    recurrency_rule = rr.rrule(
+        freq=freq,
+        interval=interval,
+        dtstart=start,
+        until=end,
+        cache=True,
+    )
+
+    return recurrency_rule
+
+
+def convert_to_rrule_freq(string: str) -> tuple[int, int]:
+    """
+    Convert a string to a rrule frequency and interval.
+
+    Args:
+        string (str): The string to be converted. Should be in the format of "1h" or "1d" or "1w".
+
+    Returns:
+        tuple[int, int]: The rrule frequency and interval.
+    """
+    freq = freq_map[string[-1]]
+    interval = int(string[:-1])
+    return freq, interval
