+++ assume/scenario/loader_csv.py
@@ -18,16 +18,10 @@ from assume.common.base import LearningConfig
 from assume.common.exceptions import AssumeException
 from assume.common.forecasts import CsvForecaster, Forecaster
 from assume.common.market_objects import MarketConfig, MarketProduct
+from assume.common.utils import convert_to_rrule_freq
 from assume.world import World
 
 logger = logging.getLogger(__name__)
-
-freq_map = {
-    "h": rr.HOURLY,
-    "m": rr.MINUTELY,
-    "d": rr.DAILY,
-    "w": rr.WEEKLY,
-}
 
 
 def load_file(
@@ -113,21 +107,6 @@ def load_file(
     except FileNotFoundError:
         logger.info(f"{file_name} not found. Returning None")
         return None
-
-
-def convert_to_rrule_freq(string: str) -> tuple[int, int]:
-    """
-    Convert a string to a rrule frequency and interval.
-
-    Args:
-        string (str): The string to be converted. Should be in the format of "1h" or "1d" or "1w".
-
-    Returns:
-        tuple[int, int]: The rrule frequency and interval.
-    """
-    freq = freq_map[string[-1]]
-    interval = int(string[:-1])
-    return freq, interval
 
 
 def replace_paths(config: dict, inputs_path: str):
