+++ examples/inputs/example_02b/config.yaml
@@ -17,7 +17,7 @@ base:
     learning_rate: 0.001
     training_episodes: 100
     episodes_collecting_initial_experience: 5
-    train_freq: 24
+    train_freq: 24h
     gradient_steps: -1
     batch_size: 256
     gamma: 0.99
