+++ examples/inputs/example_02c/config.yaml
@@ -17,7 +17,7 @@ dam:
     learning_rate: 0.0001
     training_episodes: 30
     episodes_collecting_initial_experience: 5
-    train_freq: 24
+    train_freq: 24h
     gradient_steps: -1
     batch_size: 128
     gamma: 0.99
@@ -67,7 +67,7 @@ tiny:
     learning_rate: 0.001
     training_episodes: 3
     episodes_collecting_initial_experience: 1
-    train_freq: 24
+    train_freq: 24h
     gradient_steps: -1
     batch_size: 128
     gamma: 0.99
