+++ tests/test_utils.py
@@ -14,6 +14,7 @@ from dateutil.tz import tzlocal
 from assume.common.market_objects import MarketConfig, MarketProduct
 from assume.common.utils import (
     aggregate_step_amount,
+    convert_to_rrule_freq,
     datetime2timestamp,
     get_available_products,
     get_products_index,
@@ -23,7 +24,7 @@ from assume.common.utils import (
     timestamp2datetime,
     visualize_orderbook,
 )
-from assume.scenario.loader_csv import convert_to_rrule_freq, make_market_config
+from assume.scenario.loader_csv import make_market_config
 
 from .utils import create_orderbook
